+++ plom/finish/assemble_solutions.py
@@ -96,7 +96,7 @@ def build_assemble_args(msgr, srcdir, short_name, outdir, t):
     return (outname, short_name, sid, covername, sfiles)
 
 
-def main(server=None, pwd=None):
+def main(testnum=None, server=None, pwd=None):
     if server and ":" in server:
         s, p = server.split(":")
         msgr = FinishMessenger(s, port=p)
@@ -145,13 +145,34 @@ def main(server=None, pwd=None):
         solution_args = []
         # get data for cover pages
         cover_args = []
-        for t in completedTests:
-            # check if the given test is ready for reassembly (and hence soln ready for assembly)
-            if (
-                (completedTests[t][0] is True)
-                and (completedTests[t][1] is True)
-                and (completedTests[t][2] == numberOfQuestions)
-            ):
+
+        if testnum is not None:
+            t = str(testnum)
+            try:
+                completed = completedTests[t]
+            except KeyError:
+                raise ValueError(
+                    f"Paper {t} does not exist or otherwise not ready"
+                ) from None
+            if not completed[0]:
+                raise ValueError(f"Paper {t} not scanned, cannot reassemble")
+            if not completed[1]:
+                raise ValueError(f"Paper {t} not identified, cannot reassemble")
+            if completed[2] == numberOfQuestions:
+                print(f"Note: paper {t} not fully marked but building soln anyway")
+            # append args for this test to list
+            cover_args.append(build_soln_cover_data(msgr, tmpdir, t, maxMarks))
+            solution_args.append(
+                build_assemble_args(msgr, tmpdir, shortName, outdir, t)
+            )
+        else:
+            for t, completed in tqdm(completedTests.items()):
+                # check if the given test is scanned and identified
+                if not (completed[0] and completed[1]):
+                    continue
+                # Maybe someone wants only the finished papers?
+                # if completed[2] != numberOfQuestions:
+                #     continue
                 # append args for this test to list
                 cover_args.append(build_soln_cover_data(msgr, tmpdir, t, maxMarks))
                 solution_args.append(