+++ plom/server/theServer.py
@@ -150,7 +150,6 @@ class Server:
     from .plomServer.serverMark import (
         MgetAllMax,
         MprogressCount,
-        MgetQuestionMax,
         MgetDoneTasks,
         MgetNextTask,
         MlatexFragment,